feat: editor zooming woo

+++ api/src/main/java/me/combimagnetron/sunscreen/util/helper/MathHelper.java
@@ -4,7 +4,7 @@ public class MathHelper {
 
     private MathHelper() {}
 
-    public static boolean isPowerOfTwo(float x) {
+    public static boolean isPowerOfTwo(double x) {
         if (x <= 0) {
             return false;
         }
